fix: address reviewer feedback on speaker_embedding
- Remove hardcoded float32 dtype from speaker embedding tensor creation,
  letting downstream .to(self.talker.dtype) handle conversion (P1)
- Add length validation for speaker_embedding (64-8192 range) to catch
  malformed vectors before they reach model execution (P2)

+++ vllm_omni/model_executor/models/qwen3_tts/qwen3_tts.py
@@ -961,7 +961,7 @@ class Qwen3TTSModel:
 
         # If a pre-computed speaker embedding is provided, build prompt items directly
         if speaker_embedding is not None and voice_clone_prompt is None:
-            spk_tensor = torch.tensor(speaker_embedding, dtype=torch.float32).to(self.device)
+            spk_tensor = torch.tensor(speaker_embedding).to(self.device)
             prompt_item = VoiceClonePromptItem(
                 ref_code=None,
                 ref_spk_embedding=spk_tensor,
